Use single socketpair for supervisor-spawner talk
Previously used a pipe and a socketpair. Apart from it being redundant,
it was possible to (un)intentionally hijack this internal pipe by
specifying /proc/self/fd/3 as a file destination.

It is much safer with a socketpair alone since one can't open a socket
via a link in /proc/self/fd (fails with ENXIO).