Remove sandboxed user or group name settings
Offer uid and gid settings instead.

Rationale: a typical chicken-and-egg problem. User and group names
should be resolved using user/group databases in sandboxed chroot. The
resulting uid and gid are used to configure uid/gid maps. In order to
configure chroot, uid/gid maps must be populated first and these are
write-once.

(Configured uid/gid maps are necessary to create any inode in a
filesystem mounted inside the sandbox. This is required if the target of
a bind mount doesn't exist.)